refactor: update validation and make it fail-fast (#917)
