+++ pkg/provider/provider_test.go
@@ -248,104 +248,6 @@ func TestParseAzureEnvironmentAzureStackCloud(t *testing.T) {
 	}
 }
 
-func TestValidateObjectFormat(t *testing.T) {
-	cases := []struct {
-		desc         string
-		objectFormat string
-		objectType   string
-		expectedErr  error
-	}{
-		{
-			desc:         "no object format specified",
-			objectFormat: "",
-			objectType:   "cert",
-			expectedErr:  nil,
-		},
-		{
-			desc:         "object format not valid",
-			objectFormat: "pkcs",
-			objectType:   "secret",
-			expectedErr:  fmt.Errorf("invalid objectFormat: pkcs, should be PEM or PFX"),
-		},
-		{
-			desc:         "object format PFX, but object type not secret",
-			objectFormat: "pfx",
-			objectType:   "cert",
-			expectedErr:  fmt.Errorf("PFX format only supported for objectType: secret"),
-		},
-		{
-			desc:         "object format PFX case insensitive check",
-			objectFormat: "PFX",
-			objectType:   "secret",
-			expectedErr:  nil,
-		},
-		{
-			desc:         "valid object format and type",
-			objectFormat: "pfx",
-			objectType:   "secret",
-			expectedErr:  nil,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.desc, func(t *testing.T) {
-			err := validateObjectFormat(tc.objectFormat, tc.objectType)
-			if tc.expectedErr != nil && err.Error() != tc.expectedErr.Error() || tc.expectedErr == nil && err != nil {
-				t.Fatalf("expected err: %+v, got: %+v", tc.expectedErr, err)
-			}
-		})
-	}
-}
-
-func TestValidateObjectEncoding(t *testing.T) {
-	cases := []struct {
-		desc           string
-		objectEncoding string
-		objectType     string
-		expectedErr    error
-	}{
-		{
-			desc:           "No encoding specified",
-			objectEncoding: "",
-			objectType:     "cert",
-			expectedErr:    nil,
-		},
-		{
-			desc:           "Invalid encoding specified",
-			objectEncoding: "utf-16",
-			objectType:     "secret",
-			expectedErr:    fmt.Errorf("invalid objectEncoding: utf-16, should be hex, base64 or utf-8"),
-		},
-		{
-			desc:           "Object Encoding Base64, but objectType is not secret",
-			objectEncoding: "base64",
-			objectType:     "cert",
-			expectedErr:    fmt.Errorf("objectEncoding only supported for objectType: secret"),
-		},
-		{
-			desc:           "Object Encoding case-insensitive check",
-			objectEncoding: "BasE64",
-			objectType:     "secret",
-			expectedErr:    nil,
-		},
-		{
-			desc:           "Valid ObjectEncoding and Type",
-			objectEncoding: "base64",
-			objectType:     "secret",
-			expectedErr:    nil,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.desc, func(t *testing.T) {
-			err := validateObjectEncoding(tc.objectEncoding, tc.objectType)
-			if tc.expectedErr != nil && err.Error() != tc.expectedErr.Error() || tc.expectedErr == nil && err != nil {
-				t.Fatalf("expected err: %+v, got: %+v", tc.expectedErr, err)
-			}
-		})
-	}
-}
-
 func TestGetContentBytes(t *testing.T) {
 	cases := []struct {
 		desc           string
@@ -475,49 +377,6 @@ func TestFormatKeyVaultObject(t *testing.T) {
 			formatKeyVaultObject(&tc.keyVaultObject)
 			if !reflect.DeepEqual(tc.keyVaultObject, tc.expectedKeyVaultObject) {
 				t.Fatalf("expected: %+v, but got: %+v", tc.expectedKeyVaultObject, tc.keyVaultObject)
-			}
-		})
-	}
-}
-
-func TestValidateFilePath(t *testing.T) {
-	cases := []struct {
-		desc        string
-		fileName    string
-		expectedErr error
-	}{
-		{
-			desc:        "file name is absolute path",
-			fileName:    "/secret1",
-			expectedErr: fmt.Errorf("file name must be a relative path"),
-		},
-		{
-			desc:        "file name contains '..'",
-			fileName:    "secret1/..",
-			expectedErr: fmt.Errorf("file name must not contain '..'"),
-		},
-		{
-			desc:        "file name starts with '..'",
-			fileName:    "../secret1",
-			expectedErr: fmt.Errorf("file name must not contain '..'"),
-		},
-		{
-			desc:        "file name is empty",
-			fileName:    "",
-			expectedErr: fmt.Errorf("file name must not be empty"),
-		},
-		{
-			desc:        "valid file name",
-			fileName:    "secret1",
-			expectedErr: nil,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.desc, func(t *testing.T) {
-			err := validateFileName(tc.fileName)
-			if tc.expectedErr != nil && err.Error() != tc.expectedErr.Error() || tc.expectedErr == nil && err != nil {
-				t.Fatalf("expected err: %+v, got: %+v", tc.expectedErr, err)
 			}
 		})
 	}
@@ -1049,49 +908,4 @@ func TestGetObjectVersion(t *testing.T) {
 	expectedVersion := "c55925c29c6743dcb9bb4bf091be03b0"
 	actual := getObjectVersion(id)
 	assert.Equal(t, expectedVersion, actual)
-}
-
-func TestValidateFilePermisssion(t *testing.T) {
-	cases := []struct {
-		desc                  string
-		filePermission        string
-		defaultFilePermission os.FileMode
-		isErrorExpected       bool
-	}{
-		{
-			desc:                  "valid file permission",
-			filePermission:        "0600",
-			defaultFilePermission: os.FileMode(0644),
-			isErrorExpected:       false,
-		},
-		{
-			desc:                  "empty file permission",
-			filePermission:        "",
-			defaultFilePermission: os.FileMode(0644),
-			isErrorExpected:       false,
-		},
-		{
-			desc:                  "invalid file permission",
-			filePermission:        "0900",
-			defaultFilePermission: os.FileMode(0644),
-			isErrorExpected:       true,
-		},
-		{
-			desc:                  "invalid octal number",
-			filePermission:        "900",
-			defaultFilePermission: os.FileMode(0644),
-			isErrorExpected:       true,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.desc, func(t *testing.T) {
-			_, err := validateFilePermission(tc.filePermission, tc.defaultFilePermission)
-			if tc.isErrorExpected {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-		})
-	}
 }
